Use CL_INTENSITY & CL_FLOAT For rendered frames

+++ src/framework/compositors/CLCompositor.ipp
@@ -1,7 +1,7 @@
 #ifndef CLCOMPOSITOR_IPP
 #define CLCOMPOSITOR_IPP
 
-template class CLCompositor< uint > ;
+template class CLCompositor< float > ;
 
 #endif // CLCOMPOSITOR_IPP
