+++ src/framework/renderers/CLRenderer.ipp
@@ -1,6 +1,6 @@
 #ifndef CLRENDERER_IPP
 #define CLRENDERER_IPP
 
-template class CLRenderer< uchar , uint > ;
+template class CLRenderer< uchar , float > ;
 
 #endif // CLRENDERER_IPP
