+++ src/framework/virtual/VirtualCLCompositor.ipp
@@ -1,6 +1,6 @@
 #ifndef VIRTUALCLCOMPOSITOR_IPP
 #define VIRTUALCLCOMPOSITOR_IPP
 
-template class VirtualCLCompositor< uint > ;
+template class VirtualCLCompositor< float > ;
 
 #endif // VIRTUALCLCOMPOSITOR_IPP
