+++ src/framework/virtual/VirtualCLRenderer.ipp
@@ -1,6 +1,6 @@
 #ifndef VIRTUALCLRENDERER_IPP
 #define VIRTUALCLRENDERER_IPP
 
-template class VirtualCLRenderer< uchar , uint > ;
+template class VirtualCLRenderer< uchar , float > ;
 
 #endif // VIRTUALCLRENDERER_IPP
